+++ src/cmds.ts
@@ -9,6 +9,9 @@ import { searchDocsCmd } from "./commands/searchdocs";
 import { addCommandCmd } from "./commands/addCommand";
 import { addPreferenceCmd } from "./commands/addPreference";
 import { insertImageAssetCmd } from "./commands/insertimageasset";
+import { loginCmd } from "./commands/login";
+import { publicCmd } from "./commands/publish";
+import { attachDebuggerCmd } from "./commands/attachdebugger";
 
 export function registerAllCommands(manager: ExtensionManager) {
   manager.registerCommand("lint", async () => lintCmd(manager));
@@ -21,4 +24,7 @@ export function registerAllCommands(manager: ExtensionManager) {
   manager.registerCommand("addcommand", async () => addCommandCmd(manager));
   manager.registerCommand("addpreference", async () => addPreferenceCmd(manager));
   manager.registerCommand("insertimageasset", async () => insertImageAssetCmd(manager));
+  manager.registerCommand("login", async () => loginCmd(manager));
+  manager.registerCommand("publish", async () => publicCmd(manager));
+  manager.registerCommand("attachdebugger", async () => attachDebuggerCmd(manager));
 }