+++ src/commands/publish.ts
@@ -1,0 +1,6 @@
+import { ExtensionManager } from "../manager";
+
+export async function publicCmd(manager: ExtensionManager) {
+  manager.logger.debug("start publish process");
+  manager.runNpmExec(["ray", "publish"]);
+}