+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.4
+
+- Add `Login` command
+- Add `Publish` command
+- Add `Attach Debugger` command
+
 ## 0.3
 
 - Add `Insert Image Asset`